feat: unified minikube cluster status query
